% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_fun.R
\name{&}
\alias{&}
\title{String Concatenation}
\usage{
`&`(x, y)
}
\arguments{
\item{x}{the first string}

\item{y}{the second string}
}
\value{
A character vector of concatenated strings 1 and 2
}
\description{
A very short notation for pasting two strings/character vectors together
}
