% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_fun.R
\name{data_frame_density}
\alias{data_frame_density}
\title{Calculate data.frame Density (antonym: Sparsity)}
\usage{
data_frame_density(my_df, group_by_var = NULL, var_prefix = "")
}
\arguments{
\item{my_df}{The data frame for which you want to calculate "density"}

\item{group_by_var}{an optional feature within the data frame to group by}

\item{var_prefix}{An optional prefix for the columns in the output data frame.}
}
\value{
A data frame of density values
}
\description{
This function creates a data frame with a measure of "fullness" (density) of
each feature in the data frame.  This can be done at the top-level of the
data frame or grouped by some categorical variable of interest.
}
