% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_fun.R
\name{extract_string_body}
\alias{extract_string_body}
\title{Extract a character vector from another character vector}
\usage{
extract_string_body(master_string, string_head, string_tail)
}
\arguments{
\item{master_string}{A character vector to be cut}

\item{string_head}{The pattern prefix}

\item{string_tail}{The pattern suffix}
}
\value{
A character vector the same dimensions as master_string
}
\description{
This function extracts a body of string from between a prefix-string (head) and suffix-string (tail) of a character vector.
}
