% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_fun.R
\name{left}
\alias{left}
\title{Extract characters from the start (left) of a string}
\usage{
left(my_str, nc)
}
\arguments{
\item{my_str}{A character vector to be cut}

\item{nc}{Number of characters from the left to be kept}
}
\value{
A character vector of extract strings of length nchar=nc
}
\description{
This function simply returns a number of specified characters from the start of a string/character vector.
}
