% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_fun.R
\name{ones}
\alias{ones}
\title{Returns a matrix of ones}
\usage{
ones(x, y = 1)
}
\arguments{
\item{x}{A matrix OR the x-dimensions of a matrix to be created.}

\item{y}{(optional) the y-dimensions of a matrix to be created.}
}
\value{
A matrix of same dims as x or x*y
}
\description{
This function simply returns a matrix of ones with specified dimensions.
}
