% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_fun.R
\name{pad_zeroes}
\alias{pad_zeroes}
\title{String Zero Padding}
\usage{
pad_zeroes(my_str, my_nchars)
}
\arguments{
\item{my_str}{the input number or string}

\item{my_nchars}{the desired fixed-length of the output string}
}
\value{
A string of my_str padding with zeroes to achieve a string length of my_nchars
}
\description{
An intuitive way to provide a fixed-width string by padding with leading zeroes.
}
