% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_fish_LW}
\alias{plot_fish_LW}
\title{A Fish Length-Weight Plotting Function}
\usage{
plot_fish_LW(fish_length)
}
\arguments{
\item{fish_length}{If 1, uses default fish length data set, otherwise uses a vector of fish length data}
}
\value{
Plot of fish length-weight relationship.
}
\description{
Quickly plot fish length-weight data
}
\details{
More details on how to make a slick fish length-weight plot
}
\examples{
plot_fish_LW(fish_length = 1)

}
\keyword{fish,}
\keyword{length,}
\keyword{weight}

