% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_fun.R
\name{right}
\alias{right}
\title{Extract characters from the end (right) of a string}
\usage{
right(my_str, nc)
}
\arguments{
\item{my_str}{A character vector to be cut}

\item{nc}{Number of characters from the right to be kept}
}
\value{
A character vector of extract strings of length nchar=nc
}
\description{
This function simply returns a number of specified characters from the end of a string/character vector.
}
