% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_fun.R
\name{rvals}
\alias{rvals}
\title{Returns the rvals coordinate of a matrix}
\usage{
rvals(x, y = 1, xoffset = 0, yoffset = 0)
}
\arguments{
\item{x}{A matrix OR the x-dimensions of a matrix to be created.}

\item{y}{(optional) the y-dimensions of a matrix to be created.}

\item{xoffset}{(optional) the x-offset of the origin}

\item{yoffset}{(optional) the y-offset of the origin}
}
\value{
A matrix of same dims as x or x*y
}
\description{
This function simply returns a number corresponding to the r-index of a matrix, or creates a matrix with specified dimensions.
}
