% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_fun.R
\name{vvals}
\alias{vvals}
\title{Returns the vvals of a matrix}
\usage{
vvals(x, y = 1, xoffset = 0, yoffset = 0)
}
\arguments{
\item{x}{A matrix OR the x-dimensions of a matrix to be created.}

\item{y}{(optional) the y-dimensions of a matrix to be created.}
}
\value{
A matrix of same dims as x or x*y
}
\description{
This function simply returns the element number of the matrix with specified dimensions (the pixel number for images)
}
