% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_fun.R
\name{xvals}
\alias{xvals}
\title{Returns the xval coordinate of a matrix}
\usage{
xvals(x, y = 1)
}
\arguments{
\item{x}{A matrix OR the x-dimensions of a matrix to be created.}

\item{y}{(optional) the y-dimensions of a matrix to be created.}
}
\value{
A matrix of same dims as x or x*y
}
\description{
This function simply returns a number correspondong to the x-index of a matrix, or creates a matrix with specified dimensions.
}
